requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
pydantic>=2.6.0
python-dotenv>=1.0.0
moviepy>=1.0.3
openai>=1.3.7
anthropic>=0.7.7
google-generativeai>=0.3.2
requests>=2.31.0
aiofiles>=23.2.1
pydub>=0.25.1
openai-whisper>=20231117
torch>=2.0.0
numpy>=1.21.0
aiohttp>=3.9.0
